Corrected for strange behavior
My original correction of `os.execute` was based on the fact that on my system (fedora) it returned as per the 5.2+ docs, but it is supposed to (and normally does) return as specified in the 5.1- docs. This patch works in both situations. I believe the strange behavior to be an issue with the fedora package of neovim, but it may be present elsewhere.